Dedupe verifyChecksum method
